# docs/agent-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BDO App Agent API",
    "description": "Read-only endpoints exposing SBA loan project data to the Foundry 'SBA Deal Copilot' agent. All endpoints require the x-agent-api-key header. Typical flow: call searchProjects to resolve a business or project name to a projectId, then call the per-project endpoints.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://sbaprequalifier-cbg7h5athsdkgxgm.southcentralus-01.azurewebsites.net",
      "description": "BDO App (set to the slot this agent should read from)"
    }
  ],
  "security": [{ "agentApiKey": [] }],
  "paths": {
...
    "/api/agent/sop-search": {
      "get": {
        "operationId": "searchSopPolicy",
        "summary": "Search the SBA SOP 50 10 rulebook",
        "description": "Semantic search over the full text of SBA SOP 50 10 (current edition). Use this for ANY question about SBA eligibility, program rules, franchise requirements, change-of-ownership rules, equity injection, guaranty requirements, or lender procedures. Returns the most relevant SOP passages, each labeled with its section breadcrumb (e.g. 'Section A > Chapter 3 > ...') — cite these section labels in answers. Ask focused questions; call multiple times with different phrasings for multi-part policy questions.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "A focused policy question or topic, e.g. 'franchise directory requirement for 7(a) loans' or 'seller standby note equity injection rules'"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked SOP passages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "section": { "type": "string", "description": "SOP section breadcrumb to cite" },
                          "text": { "type": "string" },
                          "score": { "type": "number" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "agentApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-agent-api-key"
      }
    }
  }
}